+++ tienda/src/main/java/ms/tienda/repository/ProveedoresRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProveedoresRepository extends JpaRepository<Proveedores,Long> {
+public interface ProveedoresRepository extends JpaRepository<Proveedores,Double> {
 }
