+++ tienda/src/main/java/ms/tienda/service/IProveedoresService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface IProveedoresService {
     public List<Proveedores> readAll();
 
-    public Proveedores readById(Long id);
+    public Proveedores readById(Double id);
 
     public Proveedores insert(Proveedores proveedor);
 
     public Proveedores update(Proveedores proveedor);
 
-    public void delete(Long id);
+    public void delete(Double id);
 
 }
